validations

+++ src/application/service/solicitacao.hora.extra.service.ts
@@ -1,11 +1,25 @@
 import SolicitacaoHoraExtraEntity, { SolicitacaoHoraExtraProps } from "../../domain/entity/solicitacao.hora.extra";
 import { SolicitacaoHoraExtraRepository } from "../../domain/repository/solicitacao.hora.extra";
+import AppError from "../errors/AppError";
 
 export default class SolicitacaoHoraExtraService {
-    constructor(
-    private readonly solicitacaoHoraExtraRepository: SolicitacaoHoraExtraRepository) {}
+    constructor(private readonly solicitacaoHoraExtraRepository: SolicitacaoHoraExtraRepository) {}
 
     async create(input: SolicitacaoHoraExtraProps) {
+
+        const camposObrigatorios: string[] = [
+            "funcionario_id",
+            "solicitante_id",
+            "data_solicitacao",
+            "horas_extras",
+        ];
+
+        for (const campo of camposObrigatorios) {
+            if (!input[campo]) {
+                throw new AppError(`${campo} obrigatório`, 400);
+            }
+        }
+
         const solicitacaoHoraExtra = new SolicitacaoHoraExtraEntity(input);
         const newSolicitacaoHoraExtra =
         await this.solicitacaoHoraExtraRepository.insert(solicitacaoHoraExtra);
@@ -20,6 +34,24 @@ export default class SolicitacaoHoraExtraService {
     }
 
     async update(id: number, input: SolicitacaoHoraExtraProps): Promise<SolicitacaoHoraExtraEntity> {
+        const camposObrigatorios: string[] = [
+            "funcionario_id",
+            "solicitante_id",
+            "data_solicitacao",
+            "horas_extras",
+        ];
+
+        for (const campo of camposObrigatorios) {
+            if (!input[campo]) {
+                throw new AppError(`${campo} obrigatório`, 400);
+            }
+        }
+
+        const solicitacaoExisting = await this.solicitacaoHoraExtraRepository.getById(id);
+        if (!solicitacaoExisting) {
+            throw new AppError("Solicitação não encontrado");
+        }
+
         const solicitacaoHoraExtra = new SolicitacaoHoraExtraEntity(input);
         const updateSolicitacaoHoraExtra =
          await this.solicitacaoHoraExtraRepository.update(id,solicitacaoHoraExtra);
@@ -29,6 +61,12 @@ export default class SolicitacaoHoraExtraService {
 
 
     async delete(id: number): Promise<void> {
+        const solicitacaoExisting =
+        await this.solicitacaoHoraExtraRepository.getById(id);
+        if (!solicitacaoExisting) {
+            throw new AppError("Solicitação não encontrado");
+        }
+
         await this.solicitacaoHoraExtraRepository.delete(id);
     }
 }
